+++ SMEngine/SMEngine.cs
@@ -9,68 +9,13 @@ using System.IO;
 using System.Linq;
 using System.Net;
 using System.Reflection;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Media.Imaging;
-using System.Net.Http;
 
 
 namespace SMEngine
 {
-
-    public class Tracker
-    {
-        private static readonly HttpClient _http = new HttpClient();
-
-        // Hard-coded configuration (no ConfigurationManager)
-        private const string Endpoint = "https://localhost:8080/track"; // TODO: replace with your endpoint
-        private const bool Enabled = false; // Set true to enable tracking
-        private const int TimeoutSeconds = 3;
-
-        private static string getEndpoint()
-        {
-            if (!Enabled) return null;
-            return string.IsNullOrWhiteSpace(Endpoint) ? null : Endpoint;
-        }
-
-        private static string JsonEscape(string s)
-        {
-            if (s == null) return string.Empty;
-            return s.Replace("\\", "\\\\").Replace("\"", "\\\"");
-        }
-
-        public void phoneHome(string appName, string host, string username)
-        {
-            var endpoint = getEndpoint();
-            if (string.IsNullOrWhiteSpace(endpoint)) return; // disabled if not configured
-
-            // Fire-and-forget
-            Task.Run(async () =>
-            {
-                try
-                {
-                    using var cts = new CancellationTokenSource(TimeSpan.FromSeconds(TimeoutSeconds));
-                    var payload = "{\"appName\":\"" + JsonEscape(appName) + "\"," +
-                                  "\"host\":\"" + JsonEscape(host) + "\"," +
-                                  "\"user\":\"" + JsonEscape(username) + "\"," +
-                                  "\"timestamp\":\"" + DateTime.UtcNow.ToString("o") + "\"}";
-                    using var content = new StringContent(payload, Encoding.UTF8, "application/json");
-                    using var req = new HttpRequestMessage(HttpMethod.Post, endpoint)
-                    {
-                        Content = content
-                    };
-                    var resp = await _http.SendAsync(req, cts.Token).ConfigureAwait(false);
-                    // no throw; just best-effort
-                }
-                catch (Exception ex)
-                {
-                    // swallow all; optionally trace in debug
-                    Debug.WriteLine($"Tracker phoneHome failed: {ex.Message}");
-                }
-            });
-        }
-    }
 
     public partial class CSMEngine
     {
@@ -250,9 +195,11 @@ namespace SMEngine
 
         private loginInfo _login;
         static Random r = new Random();
-        public CSMEngine(bool doStart)
+        public async CSMEngine(bool doStart)
         {
-            Settings = new CSettings();
+
+
+            
             _imageQueue = new Queue<ImageSet>();
             GalleryTable = new System.Data.DataTable();
             _imageDictionary = new Dictionary<string, ImageSet>();  //image id, and url
@@ -260,14 +207,17 @@ namespace SMEngine
             GalleryTable.Columns.Add(new System.Data.DataColumn("Category", typeof(string)));
             GalleryTable.Columns.Add(new System.Data.DataColumn("Album", typeof(string)));
             AllAlbums = new List<Album>();
+            Settings = new CSettings();
             Login = new loginInfo();
-            tracker.phoneHome("andyScreenSaver", Dns.GetHostName(), Environment.UserName);
+
             loadConfiguration();
             if (doStart)
             {
                 start();
             }
             setupJob();
+            var username = await Api.GetAuthenticatedUser();
+            tracker.phoneHome("andyScreenSaver", Dns.GetHostName(), username);
         }
         public CSMEngine() : this(true)
         {
@@ -368,6 +318,7 @@ namespace SMEngine
                 return Settings;
             }
         }
+        
         public loginInfo getLogin()
         {
             return Login;
