+++ SMEngine/SMEngine.csproj
@@ -151,6 +151,7 @@
     <Compile Include="loginInfo.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SMEngine.cs" />
+    <Compile Include="Tracker.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
